requirements.txt
django>=3.2.14
django-ckeditor
Pillow